# BE/openapi3.yaml
openapi: 3.0.1
info:
  title: MovieSter API
  description: Api to access MovieSter application
  license:
    name: MIT
  version: 3.5.0
...
paths:
...
  /comments:
    post:
      summary: Create new comment with an optional image
      operationId: createComment
      tags:
        - authenticated routes
        - comments
      parameters:
        - in: cookie
          name: access-token
          required: true
          schema:
            type: string
          description: JWT Access token
      requestBody:
        description: Send new comment data along with an optional image
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                description:
                  type: string
                rating:
                  type: number
                movieId:
                  type: string
                userId:
                  type: string
                image:
                  type: string
                  format: binary
              required:
                - description
                - rating
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: "^[0-9a-fA-F]{24}$"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "5XX":
          $ref: "#/components/responses/UnexpectedError"
    patch:
      summary: Update comment
      operationId: updateComment
      tags:
        - authenticated routes
        - comments
      parameters:
        - in: cookie
          name: access-token
          required: true
          schema:
            type: string
          description: JWT Access token
      requestBody:
        description: send new comment data
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CommentUpdate"
        required: true
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        5XX:
          $ref: "#/components/responses/UnexpectedError"
  /comments/{commentId}:
    get:
      summary: get comment data by id
      operationId: getCommentById
      tags:
        - authenticated routes
        - comments
      parameters:
        - in: path
          name: commentId
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
          required: true
          description: comment's id
        - in: cookie
          name: access-token
          required: true
          schema:
            type: string
          description: JWT Access token
      responses:
        "200":
          description: comment data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        5XX:
          $ref: "#/components/responses/UnexpectedError"
    delete:
      summary: delete comment data by id
      operationId: deleteCommentById
      tags:
        - authenticated routes
        - comments
      parameters:
        - in: path
          name: commentId
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
          required: true
          description: comment's id
        - in: cookie
          name: access-token
          required: true
          schema:
            type: string
          description: JWT Access token
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        5XX:
          $ref: "#/components/responses/UnexpectedError"
  /comments/user/{userId}:
    get:
      summary: get comments by user id
      operationId: getCommentsByUserId
      tags:
        - authenticated routes
        - comments
      parameters:
        - in: path
          name: userId
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
          required: true
          description: user's id
        - in: cookie
          name: access-token
          required: true
          schema:
            type: string
          description: JWT Access token
      responses:
        "200":
          description: comment array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        5XX:
          $ref: "#/components/responses/UnexpectedError"
  /comments/movie/{movieId}:
    get:
      summary: get comments by movie id
      operationId: getCommentsByMovieId
      tags:
        - authenticated routes
        - comments
      parameters:
        - in: path
          name: movieId
          schema:
            type: number
          required: true
          description: movie's id
        - in: cookie
          name: access-token
          required: true
          schema:
            type: string
          description: JWT Access token
      responses:
        "200":
          description: comment array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        5XX:
          $ref: "#/components/responses/UnexpectedError"
  /comments/thread/${mainCommentId}:
    get:
      summary: get comments thread by main comment id
      operationId: getCommentsThread
      tags:
        - authenticated routes
        - comments
      parameters:
        - in: path
          name: mainCommentId
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
          required: true
          description: main comment id
        - in: cookie
          name: access-token
          required: true
          schema:
            type: string
          description: JWT Access token
      responses:
        "200":
          description: comment array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        5XX:
          $ref: "#/components/responses/UnexpectedError"
...
components:
  responses:
    Unauthorized:
      description: Unauthorized request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NoContent:
      description: No content
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    UnexpectedError:
      description: Unexpected Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
...
    Comment:
      type: object
      properties:
        _id:
          type: string
          pattern: "^[0-9a-fA-F]{24}$"
        movieId:
          type: integer
        userId:
          type: string
          pattern: "^[0-9a-fA-F]{24}$"
        mainCommentId:
          type: string
          pattern: "^[0-9a-fA-F]{24}$"
        description:
          type: string
          format: uri
        rating:
          type: integer
          minimum: 0
          maximum: 10
        createdAt:
          type: string
          format: date-time
          description: "Optional creation date"
        updatedAt:
          type: string
          format: date-time
          description: "Optional update date"
      required:
        - movieId
        - userId
        - description
        - rating
...
    CommentUpdate:
      type: object
      properties:
        _id:
          type: string
          pattern: "^[0-9a-fA-F]{24}$"
        movieId:
          type: integer
        userId:
          type: string
          pattern: "^[0-9a-fA-F]{24}$"
        mainCommentId:
          type: string
          pattern: "^[0-9a-fA-F]{24}$"
        description:
          type: string
          format: uri
        rating:
          type: integer
          minimum: 0
          maximum: 10
      required:
        - _id
    CommentArray:
      type: array
      items:
        $ref: "#/components/schemas/Comment"